change the name of "Operations Center" to "Workflow Center" and move "workflow editor" to "workflow center"
