+++ src/app/header/header.component.html
@@ -6,10 +6,9 @@
     </a>
   </div>
   <div class="header-nav">
-    <a routerLink="/ui/managementCenter" routerLinkActive="active" class="nav-link"><span class="nav-text">Management Center</span></a>
-    <a routerLink="/ui/operationsCenter" routerLinkActive="active" class=" nav-link"><span class="nav-text">Operations Center</span></a>
-    <a routerLink="/ui/workflowEditor" routerLinkActive="active" class=" nav-link"><span class="nav-text">Workflow Editor</span></a>
-    <a routerLink="/ui/solutionCenter" routerLinkActive="active" class="nav-link"><span class="nav-text">Solution Center</span></a>
+    <a routerLink="ui/managementCenter" routerLinkActive="active" class="nav-link"><span class="nav-text">Management Center</span></a>
+    <a routerLink="ui/workflowCenter" routerLinkActive="active" class=" nav-link"><span class="nav-text">Workflow Center</span></a>
+    <a routerLink="ui/solutionCenter" routerLinkActive="active" class="nav-link"><span class="nav-text">Solution Center</span></a>
   </div>
   <div class="header-actions">
     <div class="active nav-link nav-icon" (click)="openSetting=true" style="cursor:pointer" >
