--- src/app/operations-center/operations-center.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-operations-center',
-  templateUrl: './operations-center.component.html',
-  styleUrls: ['./operations-center.component.scss']
-})
-export class OperationsCenterComponent implements OnInit {
-
-  constructor() {
-  }
-
-  ngOnInit() {
-  }
-
-}
