+++ src/app/workflow-center/active-workflow/active-workflow.component.scss
