+++ src/app/workflow-center/history-workflow/history-workflow.component.scss
