+++ src/app/workflow-center/workflow-center-routing.module.ts
@@ -1,33 +1,35 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { OperationsCenterComponent } from './operations-center.component';
+import { WorkflowCenterComponent } from './workflow-center.component';
 import { ActiveWorkflowComponent } from './active-workflow/active-workflow.component';
 import { HistoryWorkflowComponent } from  './history-workflow/history-workflow.component';
 import { WorkflowViewerComponent } from './workflow-viewer/workflow-viewer.component';
 import { RunWorkflowComponent } from './run-workflow/run-workflow.component';
+import { WorkflowEditorComponent } from './workflow-editor/workflow-editor.component';
 
-const OperationsCenterRoutes: Routes = [
+const WorkflowCenterRoutes: Routes = [
   {
     path: '',
-    component: OperationsCenterComponent,
+    component: WorkflowCenterComponent,
     children: [
       {path: '', redirectTo: 'activeWorkflow'},
       {path: 'activeWorkflow', component: ActiveWorkflowComponent},
       {path: 'historyWorkflow', component: HistoryWorkflowComponent},
       {path: 'workflowViewer', component: WorkflowViewerComponent},
       {path: 'runWorkflow', component: RunWorkflowComponent},
+      {path: 'workflowEditor', component: WorkflowEditorComponent},
     ]
   }
 ];
 
 @NgModule({
   imports: [
-    RouterModule.forChild(OperationsCenterRoutes)
+    RouterModule.forChild(WorkflowCenterRoutes)
   ],
   exports: [
     RouterModule
   ]
 })
-export class OperationsCenterRoutingModule {
+export class WorkflowCenterRoutingModule {
 }
