+++ src/app/workflow-center/workflow-center.component.html
@@ -9,14 +9,19 @@
       <clr-icon clrVerticalNavIcon shape="history"></clr-icon>
       History Workflow
     </a>
-    <a clrVerticalNavLink routerLink="./workflowViewer" routerLinkActive="active">
-      <clr-icon clrVerticalNavIcon shape="view-cards"></clr-icon>
-      Workflow Viewer
-    </a>
     <a clrVerticalNavLink routerLink="./runWorkflow" routerLinkActive="active">
       <clr-icon clrVerticalNavIcon shape="play"></clr-icon>
       Run Workflow
     </a>
+    <a clrVerticalNavLink routerLink="./workflowEditor" routerLinkActive="active">
+      <clr-icon clrVerticalNavIcon shape="note"></clr-icon>
+      Workflow Editor
+    </a>
+    <a clrVerticalNavLink routerLink="./workflowViewer" routerLinkActive="active">
+      <clr-icon clrVerticalNavIcon shape="view-cards"></clr-icon>
+      Workflow Viewer
+    </a>
+
 
   </clr-vertical-nav>
   <div class="content-area">
