+++ src/app/workflow-center/workflow-center.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-workflow-center',
+  templateUrl: './workflow-center.component.html',
+  styleUrls: ['./workflow-center.component.scss']
+})
+export class WorkflowCenterComponent implements OnInit {
+  constructor() {
+  }
+
+  ngOnInit() {
+  }
+
+}
