+++ src/app/workflow-center/workflow-editor/workflow-editor.component.html
@@ -1,4 +1,4 @@
-<div class="app-workflow-canvas">
+<div class="workflow-editor">
   <div class="workflow-json-operation clear">
     <div class="workflow-editor">
       <button class="btn btn-icon" (click)="saveConfirm()" title="Save">
@@ -21,8 +21,9 @@
           <clr-icon shape="arrow left"></clr-icon>
         </button>
         <div class="editor-dropdown">
-          <dropdown-group [data]="workflowStore" [columns]="columns" [placeholders]="placeholders" [fields]="fields" needSearchIcon=true
-            (selected)="onSelected($event)" (cleared)="onRefresh()">
+          <dropdown-group [data]="workflowStore" [columns]="columns" [placeholders]="placeholders" [fields]="fields"
+                          needSearchIcon=true
+                          (selected)="onSelected($event)" (cleared)="onRefresh()">
           </dropdown-group>
         </div>
       </div>
