+++ src/app/workflow-center/workflow-editor/workflow-editor.component.scss
@@ -1,5 +1,4 @@
-.app-workflow-canvas {
-  margin: 15px;
+.workflow-editor {
   .float-right {
     float: right;
   }
@@ -7,11 +6,11 @@
     padding-bottom: 45px;
     display: inline-flex;
     width: 100%;
+    p {
+      float: right;
+    }
     .workflow-editor {
       flex: 3;
-      p {
-        float: right;
-      }
     }
     .editor {
       margin-left: 15px;
@@ -24,6 +23,7 @@
       }
       .filter-line {
         margin-left: 0px;
+        display: -webkit-inline-box;
         .editor-dropdown {
           margin-top: 18px;
           width: 85%;
@@ -31,5 +31,7 @@
       }
     }
   }
+  div.jsoneditor jsoneditor-mode-code {
+    padding: 1px !important;
+  }
 }
-
