+++ src/app/workflow-center/workflow-editor/workflow-editor.component.spec.ts
@@ -18,5 +18,4 @@ describe('WorkflowEditorComponent', () => {
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
-
 });
