+++ src/app/workflow-center/workflow-viewer/workflow-viewer.component.scss
