--- src/app/workflow-editor/workflow-canvas/workflow-canvas.component.spec.ts
@@ -1,22 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { WorkflowCanvasComponent } from './workflow-canvas.component';
-
-describe('WorkflowCanvasComponent', () => {
-  let component: WorkflowCanvasComponent;
-  let fixture: ComponentFixture<WorkflowCanvasComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ WorkflowCanvasComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(WorkflowCanvasComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-});
