--- src/app/workflow-editor/workflow-editor-routing.module.ts
@@ -1,19 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { WorkflowEditorComponent } from './workflow-editor.component';
-
-const WorkflowEditorRoutes: Routes = [
-  {path: '', component: WorkflowEditorComponent}
-];
-
-@NgModule({
-  imports: [
-    RouterModule.forChild(WorkflowEditorRoutes)
-  ],
-  exports: [
-    RouterModule
-  ]
-})
-export class WorkflowEditorRoutingModule {
-}
