--- src/app/workflow-editor/workflow-editor.component.html
@@ -1,3 +1,0 @@
-<div class="content-container offset-header">
-  <app-workflow-canvas style="width: 100%"></app-workflow-canvas>
-</div>
